refactor(data-model): rename `data-uri-codec.ts` to `codec-data-uri.ts`
The `data:` URI codec module now carries the `codec-<thing>` name its
siblings `codec-common`, `codec-json`, and `codec-realm` use. The export
map key follows (`./codec-data-uri`), as do every importer, the test
file, and the live feature document. Two import runs that the new sort
position left out of order are re-sorted.
