+++ packages/data-model/deno.jsonc
@@ -43,10 +43,10 @@
     "./api": "./src/api.ts",
     "./cell-rep": "./src/cell-rep.ts",
     "./codec-common": "./src/codec-common/index.ts",
+    "./codec-data-uri": "./src/codec-data-uri.ts",
     "./codec-json": "./src/codec-json/index.ts",
     "./codec-realm": "./src/codec-realm/index.ts",
     "./codecs": "./src/codecs.ts",
-    "./data-uri-codec": "./src/data-uri-codec.ts",
     "./fabric-bases": "./src/fabric-bases/index.ts",
     "./fabric-instances": "./src/fabric-instances/index.ts",
     "./fabric-primitives": "./src/fabric-primitives/index.ts",
