+++ packages/data-model/src/codec-data-uri.ts
