+++ packages/runner/src/uri-utils.ts
@@ -3,7 +3,7 @@ import {
   entityRefToString,
   isEntityRef,
 } from "@commonfabric/data-model/cell-rep";
-import { hasDataUriScheme } from "@commonfabric/data-model/data-uri-codec";
+import { hasDataUriScheme } from "@commonfabric/data-model/codec-data-uri";
 import { FabricHash } from "@commonfabric/data-model/fabric-primitives";
 
 import {
