+++ packages/runner/test/cell-arrays.test.ts
@@ -3,7 +3,7 @@
 import { expect } from "@std/expect";
 import { afterEach, beforeEach, describe, it } from "@std/testing/bdd";
 
-import { DATA_URI_MEDIA_TYPE } from "@commonfabric/data-model/data-uri-codec";
+import { DATA_URI_MEDIA_TYPE } from "@commonfabric/data-model/codec-data-uri";
 
 import "@commonfabric/utils/equal-ignoring-symbols";
