+++ packages/runner/test/traverse-replay/replay.ts
@@ -19,7 +19,7 @@ import {
   type FabricValue,
   hashStringOf,
 } from "@commonfabric/data-model";
-import { hasDataUriScheme } from "@commonfabric/data-model/data-uri-codec";
+import { hasDataUriScheme } from "@commonfabric/data-model/codec-data-uri";
 
 import { ExtendedStorageTransaction } from "../../src/storage/extended-storage-transaction.ts";
 import type {
